Enable admin to add/remove tracks from albums and update playback control functionality

+++ src/server/actions/album.ts
@@ -5,6 +5,7 @@ import { prisma } from "@/server/db";
 import { deleteFile } from "@/lib/supabaseBuckets";
 import { addAlbumSchema, AddAlbumSchema } from "@/lib/validations/album";
 import { getAlbumById, getAlbumByName } from "@/server/data/album";
+import { getTracksByAlbumId } from "@/server/data/track";
 
 export const addAlbum = async (values: AddAlbumSchema) => {
   const validatedFields = addAlbumSchema.safeParse(values);
@@ -37,8 +38,52 @@ export const deleteAlbum = async (id: string) => {
   const albumImageUrl = existingAlbum.image;
   const artistId = existingAlbum.artist_id;
 
-  await prisma.album.delete({
-    where: { id },
+
+  await prisma.$transaction(async (prisma) => {
+    const tracks = await getTracksByAlbumId(id);
+
+    if (!tracks) {
+      return { error: "Album has tracks associated with it!" };
+    }
+
+    const trackImageUrls = tracks.map((track) => track.image);
+    const trackAudioUrls = tracks.map((track) => track.audio_url);
+
+    await prisma.track.deleteMany({
+      where: { album_id: id },
+    });
+
+    await prisma.album.delete({
+      where: { id },
+    });
+
+    for (const trackImageUrl of trackImageUrls) {
+      if (trackImageUrl) {
+        const { error: removeAlbumImageError } = await deleteFile(
+          trackImageUrl,
+          "artists"
+        );
+        if (removeAlbumImageError) {
+          console.error(
+            `Track image deletion failed: ${removeAlbumImageError}`
+          );
+        }
+      }
+    }
+
+    for (const trackAudioUrl of trackAudioUrls) {
+      if (trackAudioUrl) {
+        const { error: removeTrackAudioError } = await deleteFile(
+          trackAudioUrl,
+          "artists"
+        );
+        if (removeTrackAudioError) {
+          console.error(
+            `Track audio deletion failed: ${removeTrackAudioError}`
+          );
+        }
+      }
+    }
   });
 
   if (albumImageUrl) {
